Updated Prayer Card - Refactored to single use EditablePrayerCard
